Fix TypeError in getGeneNum method
This error is raised by python 3 version because division always
always return a float

+++ util/Individual.py
@@ -20,7 +20,7 @@ class Individual:
 		"""
 		Get the number of genes in the genotype
 		"""
-		return len(self.genotype)/self.geneLen
+		return int(len(self.genotype)/self.geneLen)
 
 	def getGene(self, i):
 		"""
